+++ qjazz-processes/qjazz_processes/worker/prelude.py
@@ -4,6 +4,7 @@ from .exceptions import (  # noqa F401
     ProjectRequired,
 )
 from .worker import (  # noqa F401
+    PROCESS_EXECUTE_TASK,
     ProcessCacheProtocol,
     QgisJob,
     QgisProcessJob,