+++ qjazz-processes/qjazz_processes/worker/worker.py
@@ -11,14 +11,18 @@ from typing import (
     Callable,
     Iterable,
     Iterator,
+    Mapping,
     Optional,
     Sequence,
     cast,
 )
 
+import celery
 import redis
 
 from celery.signals import (
+    task_postrun,
+    task_prerun,
     worker_before_create_process,
     worker_ready,
     worker_shutdown,
@@ -34,7 +38,9 @@ from qjazz_contrib.core.celery import Job, Worker
 from qjazz_contrib.core.condition import assert_precondition
 from qjazz_contrib.core.utils import to_utc_datetime
 
+from ..callbacks import Callbacks
 from ..processing.config import ProcessingConfig
+from ..schemas import JobResults, Subscriber
 from . import registry
 from .cache import ProcessCacheProtocol
 from .config import load_configuration
@@ -51,6 +57,9 @@ from .threads import Event, PeriodicTask
 from .watch import WatchFile
 
 LinkSequence: TypeAdapter[Sequence[Link]] = TypeAdapter(Sequence[Link])
+
+
+PROCESS_EXECUTE_TASK = "process_execute"
 
 
 FILE_LINKS = "links.json"
@@ -127,6 +136,7 @@ def presence(_state) -> dict:
         qgis_version_info=Qgis.versionInt(),
         versions=QgsCommandLineUtils.allVersions(),
         result_expires=app.conf.result_expires,
+        callbacks=list(app.processes_callbacks.schemes),
     ).model_dump()
 
 
@@ -211,6 +221,8 @@ class QgisWorker(Worker):
 
         assert_precondition(not hasattr(QgisWorker, "_storage"))
         QgisWorker._storage = conf.storage.create_instance()
+
+        self._processes_callbacks = Callbacks(conf.callbacks)
 
         #
         # Init cleanup task
@@ -425,6 +437,69 @@ class QgisWorker(Worker):
 
     def create_processes_cache(self) -> Optional[ProcessCacheProtocol]:
         return None
+
+    @property
+    def processes_callbacks(self) -> Callbacks:
+        return self._processes_callbacks
+
+
+#
+# Callbacks
+#
+
+MaybeSubscriber: TypeAdapter = TypeAdapter(Subscriber | None)
+
+
+@task_prerun.connect
+def on_task_prerun(
+    sender: object,
+    task_id: str,
+    task: celery.Task,
+    args: Sequence,
+    kwargs: Mapping,
+    **_,
+):
+    if task.name != f"{task.app.service_name}.{PROCESS_EXECUTE_TASK}":
+        return
+
+    subscriber = MaybeSubscriber.validate_python(kwargs["__run_config__"]["request"].get("subscriber"))
+    if subscriber and subscriber.in_progress_uri:
+        task.app.processes_callbacks.in_progress(
+            str(subscriber.in_progress_uri),
+            task_id,
+        )
+
+
+@task_postrun.connect
+def on_task_postrun(
+    sender: object,
+    task_id: str,
+    task: celery.Task,
+    args: Sequence,
+    kwargs: Mapping,
+    retval: JobResults,
+    state: str,
+    **_,
+):
+    if task.name != f"{task.app.service_name}.{PROCESS_EXECUTE_TASK}":
+        return
+
+    subscriber = MaybeSubscriber.validate_python(kwargs["request"].get("subscriber"))
+    if not subscriber:
+        return
+
+    match state:
+        case celery.states.SUCCESS if subscriber.success_uri:
+            task.app.processes_callbacks.on_success(
+                str(subscriber.success_uri),
+                task_id,
+                retval,
+            )
+        case celery.states.FAILURE if subscriber.failed_uri:
+            task.app.processes_callbacks.on_failure(
+                str(subscriber.failed_uri),
+                task_id,
+            )
 
 
 #